+++ zhuoan-servers/servers-biz/src/main/java/com/zhuoan/biz/event/sss/SSSGameEventDealNew.java
@@ -553,7 +553,7 @@ public class SSSGameEventDealNew {
                 // 移除数据
                 for (int i = 0; i < room.getUserIdList().size(); i++) {
                     if (room.getUserIdList().get(i)==room.getPlayerMap().get(account).getId()) {
-                        room.getUserIdList().set(i, (long)0);
+                        room.getUserIdList().set(i, 0L);
                         break;
                     }
                 }
